+++ lib/coolify_ex/client_behaviour.ex
@@ -1,6 +1,7 @@
 defmodule CoolifyEx.ClientBehaviour do
   @moduledoc false
 
+  alias CoolifyEx.ApplicationLogs
   alias CoolifyEx.Deployment
 
   @callback start_deployment(String.t(), String.t(), String.t(), keyword()) ::
@@ -8,4 +9,7 @@ defmodule CoolifyEx.ClientBehaviour do
 
   @callback fetch_deployment(String.t(), String.t(), String.t()) ::
               {:ok, Deployment.t()} | {:error, term()}
+
+  @callback fetch_application_logs(String.t(), String.t(), String.t(), keyword()) ::
+              {:ok, ApplicationLogs.t()} | {:error, term()}
 end